src/device/requirements.txt
 azure-iot-device
 paho-mqtt
 pyzbar